requirements2.txt
addict==2.4.0
affine==2.4.0
certifi @ file:///croot/certifi_1707229174982/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.2.0
cycler==0.12.1
DCNv3==1.0
filelock @ file:///croot/filelock_1700591183607/work
flatbuffers==23.5.26
fonttools==4.44.0
ftfy==6.1.3
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
h5py==3.10.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.6
importlib-metadata==7.0.1
Jinja2 @ file:///croot/jinja2_1706733616596/work
keras==2.14.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.8.1
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
ml-dtypes==0.2.0
-e git+https://github.com/open-mmlab/mmcv.git@d9e10e11846d911e8354cd024967d3a17a88083c#egg=mmcv
mmengine==0.10.3
-e git+https://github.com/vegarddale/mmsegmentation.git@98f640fc75f0e7ef87a3238cdecbc7cacf8b2174#egg=mmsegmentation
mpmath @ file:///croot/mpmath_1690848262763/work
networkx==3.2.1
ninja==1.11.1.1
numpy @ file:///croot/numpy_and_numpy_base_1704311704800/work/dist/numpy-1.26.3-cp310-cp310-linux_x86_64.whl#sha256=a281f24b826e51f1c25bdd24960ab44b4bc294c65d81560441ba7fffd8ddd2a7
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.16.5
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pillow @ file:///croot/pillow_1707233021655/work
platformdirs==4.2.0
prettytable==3.9.0
psutil==5.9.8
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML @ file:///croot/pyyaml_1698096049011/work
rasterio==1.3.9
regex==2023.12.25
requests @ file:///croot/requests_1707355572290/work
rich==13.7.0
scikit-image==0.22.0
scipy==1.12.0
six==1.16.0
snuggs==1.4.7
sympy @ file:///croot/sympy_1701397643339/work
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorrt==8.5.3.1
termcolor==2.3.0
tifffile==2023.9.26
tomli==2.0.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
triton==2.2.0
typing_extensions @ file:///croot/typing_extensions_1705599297034/work
urllib3 @ file:///croot/urllib3_1700840488771/work
wcwidth==0.2.13
wrapt==1.14.1
yapf==0.40.2
zipp==3.17.0
